requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-rest-framework==0.1.0
django==1.11
djangorestframework==3.7.0  # via django-rest-framework
olefile==0.44             # via pillow
pillow==4.3.0
pytz==2017.2              # via django
